requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
PyMuPDF
python-dotenv
googletrans